Allow to pass custom configurations to LSP servers
Some language servers may require additional configuration. The main
case I found for this is to change intelephense settings so that it
dumps its weird $HOME/intelephense empty directory to a different
location instead of in my home directory.

To make this change, I replaced the numeric table of servers so that
rather than strings, it iterates over additional tables, which may
enclose an optional parameter defining server configs.

This is good because now I can change the settings for a specific server
as well. For instance, this may make easier to configure for me the
language server for UnoCSS in the upcoming future.