Narrowings and maybe some other things
twilight provides narrowings, which allows to conceal some parts of the
code that are not focused, so that they don't distract, which is useful
for streaming.

However, it seems that twilight is only compatible with themes that
support tree-sitter, and none of the default themes in Neovim 0.9
support it (expected). Until I can confirm whether the classic Vim theme
for Neovim 0.10 supports tree-sitter, I'm just going to throw the
default VSCode theme, because at least it is widely known and I won't
have to give many explanations on what I use.